mrpipeline: Functions to set up a pipeline for performing Mendelian randomization

Pipeline to perform forward mendelian randomization analyses, including multiple comparison assessments and colocalization analyses. This package provides manuscript ready tables and figures.

Author

Maintainer: Benjamin Zuckerman benjamin.zuckerman@kcl.ac.uk (ORCID)

Authors:

Alasdair Warwick (ORCID)